reorganize docs, create data_manip
