--- docs/source/data_wrangler.rst
@@ -1,10 +1,0 @@
-
-Data wrangler
-=============
-
-Data manipulation, interfacing with the database and local file, to get clean dataframes.
-
-.. automodule:: dreem_nap.data_wrangler
-   :members:
-   :undoc-members:
-   :show-inheritance:
