--- docs/source/database.rst
@@ -1,12 +1,0 @@
-
-
-Database
-========
-
-In this section, will be described the database manipulation functions.
-
-.. automodule:: dreem_nap.database
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
