+++ docs/source/gallery.rst
@@ -1,0 +1,2 @@
+Example gallery
+===============
