--- docs/source/getting_started.rst
@@ -1,14 +1,0 @@
-Getting started
-===============
-
- 
-
-With Routines
--------------
-
-
-
-With Jupyter Notebooks
-----------------------
-
-
