--- docs/source/plot.rst
@@ -1,20 +1,0 @@
-Plots
-=====
-
-In this section, will be presented which plots you can use to visualize DREEM's output.
-
-.. note:: 
-
-   Most plots requires a Pandas dataframe, a specific sample or a study, and sometimes a (list of) construct(s).
-
-.. note:: 
-
-   A study is a series of tubes that are relevant to be studied together (replicates, change of only one experimental condition, etc).
-
-
-.. automodule:: dreem_nap.plot
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-
