+++ docs/source/release_notes.rst
@@ -1,0 +1,2 @@
+Release_notes
+=============
