--- docs/source/study.rst
@@ -1,18 +1,0 @@
-Study class
-===========
-
-In this section, will be presented the Study class. 
-A study is a series of samples that are relevant to be studied together (replicates, change of only one experimental condition, etc).
-Study class stores relevant information about the study, such as the samples names, or what experimental condition changes in the study and how for each sample.
-
-.. note:: 
-
-   The study class is used as an input for many NAP functions.
-
-
-.. automodule:: dreem_nap.study
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-
