--- docs/source/utils.rst
@@ -1,7 +1,0 @@
-Utils
-=====
-
-.. automodule:: dreem_nap.utils
-   :members:
-   :undoc-members:
-   :show-inheritance:
