+++ make_docs.sh
@@ -1,3 +1,9 @@
+
+echo ''
+echo '         PIP INSTALL MODULE         '
+echo '------------------------------------'
+pip install .
+cd docs
 echo ''
 echo '            MAKE  CLEAN              '
 echo '------------------------------------'
